requirements.txt
replicate>=0.20.0
python-dotenv>=1.0.0
pytest==7.4.2
requests>=2.31.0
pytest-xdist==3.5.0
pytest-timeout==2.3.1
click>=8.1.7
tiktoken>=0.5.2
rich>=13.7.0
